Added Rate Trainer to the mod menu
