+++ core/src/bms/player/beatoraja/pattern/PatternModifier.java
@@ -1,12 +1,9 @@
 package bms.player.beatoraja.pattern;
 
-import java.util.Arrays;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.logging.Logger;
 
-import bms.player.beatoraja.RandomTrainer;
 import com.badlogic.gdx.utils.IntArray;
 
 import bms.model.BMSModel;
